\name{addColumnTotals}
\alias{addColumnTotals}
\title{Add a column totals row, preserving meta and names.}
\usage{
  addColumnTotals(x)
}
\description{
  Add a column totals row, preserving meta and names.
}
\examples{
x <- runs.mean.mean$all$gpmorb
x <- env.base$years1_5$results$quantiles$kids
x <- years1_5$runs$means$all$gpmorb
x <- results$means$all$kids
addColumnTotals(x)
}

