% Generated by roxygen2 (4.0.1): do not edit by hand
\name{append.lists}
\alias{append.lists}
\title{Append the elements of list.src into the corresponding elements of list.dest.}
\usage{
append.lists(list.dest, list.src, by.name = FALSE, flatten.src = FALSE)
}
\arguments{
\item{list.dest}{null or a list of nulls, in which case list.src is
 output as a list of lists
 a list of elements, or lists, in which case
 the elements of list.src are appendend to the
 corresponding elements in list.dest}

\item{list.src}{a list of any element type. Element type may also be a list.
 Each element is appended to the corresponding element of list.dest.}

\item{by.name}{If FALSE (default) merge so that
 list.dest[[1]] is a list of list.dest[[1]] and list.src[[1]],
 list.dest[[2]] is a list of list.dest[[2]] and list.src[[2]], etc...
 If TRUE the lists are merged by name rather than by position.
 When merging by name lists do not have to have the same
 number of elements and named elements that appear in
 only one of the lists will simply be appended to the result.}

\item{flatten.src}{If FALSE then elements E of list.src will be added as is. If E is a list
 it will be added as a list.
 If TRUE then if E is a list its contents will be added, rather than
 as a list.}
}
\value{
a list of lists. The inner list contains the
 elements supplied by list.src appended to the existing
 elements of list.dest.
}
\description{
NB: attributes of list.dest are replaced by list.src.
}
\examples{
#list.dest <- results.list; list.src <- results
#list.dest <- years1_5$runstats.collated$means
#list.src <- years6_13$runstats.collated$means
#by.name = TRUE
#r <- append.lists(list.dest, list.src, by.name)

#Test first three cases:

#ml0 <- namedList("msmoke", "fsmoke")
#x <- env.base$years1_5$outcomes; indices=names(ml0); simplify = FALSE; .FUN=table_mx_cols
#list.src <- lapply.subset(x, indices, .FUN)

#ml0 <- namedList("gptotvis", "hadmtot")
#x <- env.base$years1_5$outcomes; indices=names(ml0); simplify = FALSE; .FUN=wtdtablecols
#list.src <- lapply.subset(x, indices, .FUN)

#ml1 <- append.lists(ml0, list.src)
#ml2 <- append.lists(ml1, list.src)
#ml3 <- append.lists(ml2, list.src)

#list.dest <- ml0
#list.dest <- ml1

#Test with lists with differently named items.

list.dest <- list(foo=matrix(1:4)) # list.dest <- years1_5$run_results_collated$quantiles
list.src <- list(bar=matrix(5:10)) # list.src <- years6_13$run_results_collated$quantiles

by.name = TRUE; flatten.src = TRUE
list.result <- append.lists(list.dest , list.src, by.name = TRUE, flatten.src = TRUE)
}

