% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/support.r
\name{as.csv.string.list}
\alias{as.csv.string.list}
\title{Apply to a list using the names from the list}
\usage{
as.csv.string.list(xlist, row.names = T)
}
\arguments{
\item{xlist}{list}

\item{row.names}{either a logical value indicating whether the row names of x are to be written along with x,
 or a character vector of row names to be written}
}
\value{
a list of character vectors
}
\description{
Apply to a list using the names from the list
}
\examples{
x1 <- matrix(1:8,nrow=2)
x2 <- matrix(2:9,nrow=2)
xlist <- list(x1, x2)
names(xlist) <- c("x1", "x2")
as.csv.string.list(xlist)
}

