\name{as.data.frame.list.as.is}
\alias{as.data.frame.list.as.is}
\title{Convert list to dataframe simply by setting class attr
This preserves any matrices in the list.}
\usage{
  as.data.frame.list.as.is(xlist, row.names = NULL)
}
\arguments{
  \item{xlist}{list}

  \item{row.names}{row names for the data.frame}
}
\description{
  Convert list to dataframe simply by setting class attr
  This preserves any matrices in the list.
}

