% Generated by roxygen2 (4.0.1): do not edit by hand
\name{as_array_list_mx}
\alias{as_array_list_mx}
\title{Convert a list of matrices into an array, with each matrix in the z dimension.}
\usage{
as_array_list_mx(listmx)
}
\arguments{
\item{listmx}{list of matrices. Each must matrix the same dimensions.}
}
\value{
array with dimnames and meta from listmx
}
\description{
Convert a list of matrices into an array, with each matrix in the z dimension.
}
\examples{
listmx <- list(A=structure(matrix(c(1:6), nrow=2, dimnames=(list(1:2, 1:3))), meta=c("grpby.tag"="r1stchildethn")), B=structure( matrix(c(21:26), nrow=2, dimnames=(list(1:2, 1:3))) , meta=c("grpby.tag"="r1stchildethn")))
 as_array_list_mx(listmx)
}

