\name{as_data_frame_list_as_is}
\alias{as_data_frame_list_as_is}
\title{Convert list to dataframe simply by setting class attr
This preserves any matrices in the list.}
\usage{
  as_data_frame_list_as_is(xlist, row.names = NULL)
}
\arguments{
  \item{xlist}{list}

  \item{row.names}{row names for the data.frame}
}
\description{
  Convert list to dataframe simply by setting class attr
  This preserves any matrices in the list.
}

