% Generated by roxygen2 (4.0.1): do not edit by hand
\name{binary.levels.split}
\alias{binary.levels.split}
\title{Create a list of binary levels from factors.
Compares a vector x against a set of factors and returns
a list for each factor which indicates whether x
is equal to that factor or not.}
\usage{
binary.levels.split(x, f = sort(unique(x)))
}
\arguments{
\item{x}{vector}

\item{f}{factors. Defaults to the unique set of values in x.}
}
\value{
list of binary levels
}
\description{
Create a list of binary levels from factors.
Compares a vector x against a set of factors and returns
a list for each factor which indicates whether x
is equal to that factor or not.
}
\examples{
x <- c(1,2,3,2,1)
f <- c(1,2,3)

binary.levels.split (x, f)

\dontrun{
$`1`
[1] 1 0 0 0 1

$`2`
[1] 0 1 0 1 0

$`3`
[1] 0 0 1 0 0


x <- adjcatvar
f=sort(unique(x))
binary.levels.split(x)
}
}

