% Generated by roxygen2 (4.0.1): do not edit by hand
\name{change.cat}
\alias{change.cat}
\title{Change values in a categorical vector to a lower or higher value}
\usage{
change.cat(num, rank.col, i, new.all.dat, n.change)
}
\arguments{
\item{num}{the number of categories above you want to steal from
if you are giving and not stealing, you always give to the category
directly above  and num=1}

\item{rank.col}{the colum in which the rankings will be}

\item{i}{the interation you are up to in the while/for loop}

\item{new.all.dat}{the data you are working on and want modified}

\item{n.change}{the number of observations that need moving into or out of each
  category to get the requested proportions}
}
\value{
the categorical vector after modified
}
\description{
Change values in a categorical vector to a lower or higher value
}
\examples{
\dontrun{}
}
\seealso{
This function is called by the function \code{\link{modifyProps}}
}

