\name{chart.2series.bar.err.closure}
\alias{chart.2series.bar.err.closure}
\title{Close over parameters to chart.2series.bar.err}
\usage{
  chart.2series.bar.err.closure(title, xlab, ylab,
    result.row.base, result.row.scenario, scenario.name)
}
\description{
  Close over parameters to chart.2series.bar.err
}

