% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simframe.r
\name{checkModelVars}
\alias{checkModelVars}
\title{Checks that the variables used in all of the models
exist in simvalues, if not generate stop}
\usage{
checkModelVars(models, simframe)
}
\arguments{
\item{models}{a list of glm objects}

\item{simframe}{the simframe}
}
\value{
nothing if successful, otherwise errors
}
\description{
Checks that the variables used in all of the models
exist in simvalues, if not generate stop
}
\examples{
#simframe <- simframe.master
#checkModelVars(models, simframe)
}

