% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Collate.r
\name{collator_list_mx}
\alias{collator_list_mx}
\title{Collate and average a list of matrices.}
\usage{
collator_list_mx(runs, CI = TRUE, ...)
}
\arguments{
\item{runs}{a list of matrices, one matrix per run. Each matrix must have
 the same dimensions.}

\item{CI}{if TRUE and length(runs) > 1, lower and upper confidence intervals
are returned in additional columns}

\item{...}{Additional arguments passed to \code{\link{mean_array_z_pctile_CIs}}}
}
\value{
a matrix with the averaged values of runs.
}
\description{
Collate and average a list of matrices.
}
\examples{
run1 = structure(matrix(1:6, nrow=3, dimnames=list(1:3, c("F","M"))), meta=c(varname="earnings", grpby.tag="sex"))
run2 = structure(matrix(11:16, nrow=3, dimnames=list(1:3, c("F","M"))), meta=c(varname="earnings", grpby.tag="sex"))
runs <- list(run1=run1,run2=run2)
collator_list_mx(runs)
}

