% Generated by roxygen2 (4.0.1): do not edit by hand
\name{collator_mutiple_lists_mx}
\alias{collator_mutiple_lists_mx}
\title{Collate and average mutiple lists of matrices.}
\usage{
collator_mutiple_lists_mx(runs, CI = TRUE)
}
\arguments{
\item{runs}{a list of lists of matrices, one inner list per run.
 Each inner list may have any number of matrices,
 and each matrix may have a different sets of rows or columns.
 The matrices will be flattened into rows.}

\item{CI}{if TRUE and length(runs) > 1, lower and upper confidence intervals
 are returned in additional columns}
}
\value{
a matrix with the averaged values of runs.
}
\description{
Collate and average mutiple lists of matrices.
}
\examples{
run1_mx1 = matrix(1:2, nrow=1, dimnames=list(1, c("F","M")))
run1_mx2 = matrix(1:4, nrow=2, dimnames=list(1:2, c("F","M")), byrow = TRUE)
run1 = structure(list(run1_mx1, run1_mx2), meta=c(varname="disability_state", grpby.tag="sex"))
run2_mx1 = matrix(11:12, nrow=1, dimnames=list(1, c("F","M")))
run2_mx2 = matrix(11:14, nrow=2, dimnames=list(3:4, c("F","M")), byrow = TRUE)
run2 = structure(list(run2_mx1, run2_mx2), meta=c(varname="disability_state", grpby.tag="sex"))

runs <- list(run1=run1,run2=run2)
collator_mutiple_lists_mx(runs, CI=FALSE)
}
\keyword{internal}

