% Generated by roxygen2 (4.0.1): do not edit by hand
\name{colmeans.list}
\alias{colmeans.list}
\title{calculates the row means of each matrix
NAs are treated as zero.}
\usage{
colmeans.list(xlistm)
}
\arguments{
\item{xlistm}{a list of matrices}
}
\value{
a list
}
\description{
calculates the row means of each matrix
NAs are treated as zero.
}
\examples{
\dontrun{
colmeans.list(by.year$base$o.gptotvis)
xlistm <- by.year$base$o.gptotvis
}
}

