% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MatrixUtil.r
\name{copyMeta.list}
\alias{copyMeta.list}
\title{copy names of dimensions and meta attributes
from elements of list.src to list.dest
Called by append.list.mx().}
\usage{
copyMeta.list(list.dest, list.src)
}
\arguments{
\item{list.dest}{list of vector/matrix/arrays}

\item{list.src}{list of vector/matrix/arrays}
}
\value{
list of vector/matrix/arrays after add names and attributes
}
\description{
copy names of dimensions and meta attributes
from elements of list.src to list.dest
Called by append.list.mx().
}

