% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simframe.r
\name{createOutputMatrix}
\alias{createOutputMatrix}
\title{Create a matrix of NA with specified col/row names/lengths.}
\usage{
createOutputMatrix(varname, rows, cols)
}
\arguments{
\item{varname}{simframe source var. stored in the "varname" attribute.
 of the matrix. This is the name of the variable in the
 simframe that will be used to fill this matrix during simulation}

\item{rows}{row names, or a numeric scalar for the number of rows}

\item{cols}{columns names, or a numeric scalar for the number of cols}
}
\value{
a matrix with the "varname" attribute set to varname
}
\description{
Create a matrix of NA with specified col/row names/lengths.
}
\examples{
\dontrun{
rows <- length(children$z1msmoke1)
cols <- 5
varname <- "z1msmokeLvl1"
createOutputMatrix(varname, nrows, ncols)
}
}

