% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CategoricalAdjustment.r
\name{createSingleIterationPropensityArray}
\alias{createSingleIterationPropensityArray}
\title{Create a propensity array from a dataframe, for only a single iteration.}
\usage{
createSingleIterationPropensityArray(df, iteration_name)
}
\arguments{
\item{df}{dataframe containing n columns where n is one less than the number of categories.
This column represents the propensity to change from the 1st category to the 2nd category.}

\item{iteration_name}{used to label the single iteration in the z dim}
}
\value{
propensity array with
rows - the values for each individual micro-unit
cols - propensity to change from the 1st category to the 2nd category.
z dim - iterations/years. Only 1 iteration.
}
\description{
Create a propensity array from a dataframe, for only a single iteration.
}
\examples{
\dontrun{
df <- data.frame(year1_cat1 = 1:10/100, year1_cat2 = 11:20/100)
iteration_name <- "Year 1"
createSingleIterationPropensityArray(df, iteration_name)
}
}

