\name{detachReturn}
\alias{detachReturn}
\title{Detach an environment and return it.
NB: the returned environment contains the contents of the attached environment but is
actually a newly created different environment object from the original.}
\usage{
  detachReturn(envname)
}
\arguments{
  \item{envname}{name of the attached environment}
}
\description{
  Detach an environment and return it. NB: the returned
  environment contains the contents of the attached
  environment but is actually a newly created different
  environment object from the original.
}

