% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/support.r
\name{err}
\alias{err}
\title{Calc the 95\% error from the t Distribution.}
\usage{
err(x)
}
\arguments{
\item{x}{vector}
}
\value{
the result calculated
}
\description{
Calc the 95\% error from the t Distribution.
}
\examples{
err(x=2)
err(x=1:10)
err(x=c(4,NA))
}

