\name{evaluateLogisetExprAttribute}
\alias{evaluateLogisetExprAttribute}
\title{desired_props <- structure(1, logiset="alive & residential")
simframe <- env.base$simframe
evaluateLogisetExprAttribute(desired_props, simframe)}
\usage{
  evaluateLogisetExprAttribute(desired_props, simframe,
    varname = "")
}
\description{
  desired_props <- structure(1, logiset="alive &
  residential") simframe <- env.base$simframe
  evaluateLogisetExprAttribute(desired_props, simframe)
}

