% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MatrixUtil.r
\name{flatten_mx_to_row}
\alias{flatten_mx_to_row}
\title{Flatten a vector or matrix (r,c) into a single row matrix (1,c)
creating new column names from a combination of old rownames
and colnames.}
\usage{
flatten_mx_to_row(mx, row.names.first = FALSE)
}
\arguments{
\item{mx}{matrix}

\item{row.names.first}{if TRUE, row.names will appear first in the names of the resultant columns, eg: becomes A1,B1,A2,B2
 else row.names will appear last, eg: 1A, 1B, 2A, 2B}
}
\value{
a single row matrix (1,c)
}
\description{
A single row matrix, unlike a vector, has the advantage that
it can have a row name.
}
\details{
eg: A1, A2,
    B1, B2
}
\examples{
# 2 x 3 matrix
mx <- matrix(c(1:6), nrow = 2, ncol = 3, dimnames=list(c("A", "B"), c("1", "2", "3")))

# 1 columned matrix (2 x 1) with no col names
mx <- matrix(c(10,20), nrow = 2, ncol = 1, dimnames=list(c("0", "1"), NULL))

# 1 columned matrix (3 x 1) with no dim names
mx <- matrix(c(1,2,3), nrow = 3, ncol = 1)

# 1 columned matrix (2 x 1) with no col names and NA in row name
mx <- matrix(c(10,20), nrow = 2, ncol = 1, dimnames=list(c("0", NA), NULL))

# 1 columned matrix (1 x 1) with no col names and NA in row name
mx <- matrix(c(10), nrow = 1, ncol = 1, dimnames=list(c(NA), NULL))

# vector
mx <- c("a"=1,"b"=2,"c"=3)

# 1 columned matrix (3 x 1) with names
mx <- matrix(c(1,2,3), nrow = 3, ncol = 1, dimnames=list(c("row1", "row2", "row3"), "col1"))

flatten_mx_to_row(mx)
flatten_mx_to_row(mx, row.names.first=TRUE)
}

