% Generated by roxygen2 (4.0.1): do not edit by hand
\name{flatten_mxs_to_single_mx}
\alias{flatten_mxs_to_single_mx}
\title{Takes a list of matrices, flattens them into a single row, aligns them by column name
and then returns a single matrix.}
\usage{
flatten_mxs_to_single_mx(listmx)
}
\arguments{
\item{listmx}{list of matrices to flatten. Can have different dimensions but must have similar names
for alignment (}
}
\value{
a single matrix with
   colnames = the flattened row/col names of listmx[[1]]],
   meta = meta of listmx[[1]]
   rownames = names of listmx
}
\description{
Takes a list of matrices, flattens them into a single row, aligns them by column name
and then returns a single matrix.
}
\examples{
listmx <- list(A=structure(matrix(c(1:6), nrow=2, dimnames=(list(1:2, 1:3))), meta=c("grpby.tag"="r1stchildethn")), B=structure( matrix(c(21:26), nrow=2, dimnames=(list(1:2, 1:3))) , meta=c("grpby.tag"="r1stchildethn")))
 flatten_mxs_to_single_mx(listmx)
}
\seealso{
listmx.flat.aligned)
}

