% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getOutcomeVars}
\alias{getOutcomeVars}
\title{Get a set of outcome variables from the simframe.}
\usage{
getOutcomeVars(simframe, outcome_type_select = NULL,
  outcome_module_name = NULL, sorted = FALSE)
}
\arguments{
\item{simframe}{simframe}

\item{outcome_type_select}{a vector of outcome_types, or NULL to select all types.}

\item{outcome_module_name}{a vector of outcome_module, , or NULL to select all sets.}

\item{sorted}{return outcome variables sorted by outcome name?}
}
\value{
a vector of simframe variables (with names the same as the vector value)
}
\description{
Get a set of outcome variables from the simframe.
}
\examples{
\dontrun{
outcome_module_name = "years1_5"
outcome_module_name = c("years1_5", "years6_13")
outcome_module_name = NULL

outcome_type_select = "categorical"
outcome_type_select = "continuous"
outcome_type_select = NULL
sorted = TRUE

getOutcomeVars(simframe, outcome_type_select, outcome_module_name, sorted)
}
}

