\name{globalNamed}
\alias{globalNamed}
\title{Ssave x into global variable, ie: top frame, not just this function
using the supplied varname}
\usage{
  globalNamed(varname, x, pos = 1)
}
\arguments{
  \item{varname}{variable name}

  \item{x}{value}

  \item{pos}{environment to save into. Defaults to global
  environment.}
}
\description{
  Ssave x into global variable, ie: top frame, not just
  this function using the supplied varname
}

