% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Collate.r
\name{identify.position.last.space}
\alias{identify.position.last.space}
\title{Identify the position of the last space in each element in a character vector.
Used to identify the last space in the column names in label_flattened_mx_grping.and.CIs().
Returns a list of 2 elements.
The first element, pos.last.space.vec, is a vector identifying the position of the last space
in each element of the input vector.
The second element, num.spaces, states the number of spaces in each element of the input
vector.}
\usage{
\method{identify}{position.last.space}(col.names)
}
\arguments{
\item{col.names}{a character vector.}
}
\value{
A list with two vectors: a vector of the position of last space and
a vector of number of spaces
}
\description{
Identify the position of the last space in each element in a character vector.
Used to identify the last space in the column names in label_flattened_mx_grping.and.CIs().
Returns a list of 2 elements.
The first element, pos.last.space.vec, is a vector identifying the position of the last space
in each element of the input vector.
The second element, num.spaces, states the number of spaces in each element of the input
vector.
}
\examples{
\dontrun{
col.names <- c(" ", "1 2", "1  2", "1 2 3")
identify.position.last.space(col.names)
}
}

