% Generated by roxygen2 (4.0.1): do not edit by hand
\name{is_numeric_scalar}
\alias{is_numeric_scalar}
\title{Returns whether x is a scalar (i.e. length 1)
and numeric.}
\usage{
is_numeric_scalar(x)
}
\arguments{
\item{x}{object}
}
\value{
logical value
}
\description{
Returns whether x is a scalar (i.e. length 1)
and numeric.
}
\examples{
is_numeric_scalar(c(1,2))
is_numeric_scalar(1)
}

