% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Label.r
\name{labelTitle}
\alias{labelTitle}
\title{names the "along" dimension with "title"
eg: labelTitle(y, ROW, "Year")}
\usage{
labelTitle(xm, along, title)
}
\arguments{
\item{xm}{}

\item{along}{}

\item{title}{}
}
\description{
names the "along" dimension with "title"
eg: labelTitle(y, ROW, "Year")
}
\examples{
\dontrun{}
}

