% Generated by roxygen2 (4.0.1): do not edit by hand
\name{mean_array_z}
\alias{mean_array_z}
\title{Mean across Z dimension of a 3D array. Each row and column cell is
averaged across the Z dimension.}
\usage{
mean_array_z(xa, CI = TRUE, NA.as.zero = T, re_write_colnames = T)
}
\arguments{
\item{xa}{an array with a Z dimension}

\item{CI}{if TRUE and xa has more than 1 Z dimension, lower and upper confidence intervals
are returned in additional columns}

\item{NA.as.zero}{if TRUE (default), treat NAs as if they are zeros}
}
\value{
a matrix of means across the Z dimension. The "meta" attribute
 of xa, if any, is retained.
}
\description{
Mean across Z dimension of a 3D array. Each row and column cell is
averaged across the Z dimension.
}
\examples{
\dontrun{
xa <- env.base$years1_5$runstats$means$all$gpmorb
xa <- env.base$years6_13$runstats$means$all$burt
xa <- env.base$years1_5$runstats$means$all.by.SESBTH.2cat$gpmorb
xa <- years1_5$runstats$means$all.by.gender$gptotvis

xa <- flatten_mxlists_to_array(env.scenario$modules$years6_13$run_results$confreqs[["cond"]])

xa <- env.base$years1_5$runstats$means$all.by.ethnicity$gptotvis
xa <- env.base$years1_5$runstats$means$all.by.gender$gptotvis

result <- mean_array_z(xa)
}
}

