% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MatrixStats.r
\name{mean_array_z_pctile_CIs2}
\alias{mean_array_z_pctile_CIs2}
\title{A new version of mean_array_z_pctile_CIs().}
\usage{
mean_array_z_pctile_CIs2(xa, CI = TRUE, NA.as.zero = T,
  cat.adjustments = NULL, dict, binbreaks = NULL)
}
\arguments{
\item{xa}{an array with a Z dimension}

\item{CI}{if TRUE and xa has more than 1 Z dimension, lower and upper confidence intervals
are returned in additional columns}

\item{NA.as.zero}{if TRUE (default), treat NAs as if they are zeros}

\item{cat.adjustments}{The cat.adjustments list containing the cat.adjustments for multiple variables.
Within the function the specific cat.adjustments for the variable of interest are
extracted from the list.  Either cat.adjustments or binbreaks are needed if frequencies
of a continuous variable are being requested.}

\item{dict}{the specific project dictionary}

\item{binbreaks}{The binbreaks for the specific outcome variable.  Either binbreaks or cat.adjustments
may be provided to the function.}
}
\value{
a matrix of means
}
\description{
Differs from mean_array_z_pctile_CIs() in that the percentages
are calculated within each run and the mean of the percentages
taken as the point estimate.
}
\details{
In mean_array_z_pctile_CIs() the #' mean numerator and
denominator are taken over the runs and to #' get the point
estimate this mean numerator is divided by the mean
denominatoror to get a point estimate of the percentage.
This is equivalent to the first method (implemented in this
function) if the denominator does not change over time, but if
the denominator does change over time (e.g. this might happen #' when we are taking percentages by groups where the grouping
variable is time-variant) the results are different and the
method in this function is preferable. The same difference
between the two versions of the function is seen in the
calculation of confidence intervals.  In
mean_array_z_pctile_CIs() the 2.5th and 97.5th percentiles
of the frequencies over the runs are taken and then these are
divided by the mean denominator.  If the we have a situation as
described above where percentages are being calculated by a
time-variant grouping variable then we can end up with confidence
intervals that are greater than 100%. In this function, the
2.5th and 97.5th percentiles of the percentages are taken so
that, in the case where percentages are calculated by a
time-variant grouping variable, the confidence intervals are
more accurate and will not go above 100%. This function has
additional arguments compared to mean_array_z_pctile_CIs(),
namely, cat.adjustments, dict, and binbreaks.
}
\examples{
\dontrun{}
}

