% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MatrixStats.r
\name{mean_mx_cols_BCASO}
\alias{mean_mx_cols_BCASO}
\title{Calculates the weighted mean for each column of the matrix
Optionally subsetting and grouping by another (equal length/number of rows) variable,
which can be either a vector or a matrix.
Optionally having different weights for each column of the matrix (same length/number of rows))- wgts parameter
which can also be either a vector or a matrix.
 The group-by variable may be time-invariant or time-variant.}
\usage{
mean_mx_cols_BCASO(mx, grpby = NULL, grpby.tag = NULL, logiset = NULL,
  wgts = NULL, dict = NULL)
}
\arguments{
\item{mx}{matrix or dataframe to calculate column means of}

\item{grpby}{elements to group by, or NULL or unspecified to do no grouping.
 Same length as the columns of mx.
If the group-by variable is time-invariant grpby can be provided as a vector or as a
matrix with every column the same.  If the group-by variable is time-variable then
grpby should be a matrix with number of columns equal to number of years.}

\item{grpby.tag}{added to meta attribute of the result}

\item{logiset}{logical vector indicating which observations to include, or NULL to include all.}

\item{wgts}{elements to weight by, or NULL to do no weighting}

\item{dict}{the specific project dictionary}
}
\value{
a matrix of means. Each row is the mean for a column in mx.
 if grpby is NULL, a single column of means.
 if grpby is specified, a column of means for each group by category.
}
\description{
Calculates the weighted mean for each column of the matrix
Optionally subsetting and grouping by another (equal length/number of rows) variable,
which can be either a vector or a matrix.
Optionally having different weights for each column of the matrix (same length/number of rows))- wgts parameter
which can also be either a vector or a matrix.
 The group-by variable may be time-invariant or time-variant.
}
\examples{
mx <- matrix (c(1:10), ncol = 2) ; mx <- matrix (c(1:3, NA, 5:7, NA, 9:10), ncol = 2)
logiset = NULL
grpby = NULL ; grpby.tag = NULL
grpby <- c("A","A","A","B","B") ; grpby.tag = "AB"

mx <- matrix (c(1:4, 5, 6:7,NA,NA,10), ncol = 2)
wgts = rep(1, nrow(mx))
wgts =c(2,2,2,2,1)
grpby <- c(1,4,3,2,2) ;

wgts<-matrix(c(1,1,1,1,1,2,2,2,2,1), ncol = 2)
na.rm = FALSE ; na.rm = TRUE
}

