% Generated by roxygen2 (4.0.1): do not edit by hand
\name{merge_list_mx.by.cols}
\alias{merge_list_mx.by.cols}
\title{Takes any number of 2D matrices, each with the same number of cols but with
any set of named rows. Every col 1 of the matrices are combined into a new
matrix, as are every col 2,3 etc.}
\usage{
merge_list_mx.by.cols(...)
}
\arguments{
\item{...}{matrices, or a list of matrices}
}
\value{
matrices, or a list of matrices after merged
}
\description{
Takes any number of 2D matrices, each with the same number of cols but with
any set of named rows. Every col 1 of the matrices are combined into a new
matrix, as are every col 2,3 etc.
}
\examples{
\dontrun{
xlistm <- list(matrix(c(1:9),nrow=3,dimnames=list(c("a","b","c"),c())),matrix(c(11,15,14,18,20,21),nrow=2,dimnames=list(c("a","e"),c())),matrix(c(13,17,30,31,40,41),nrow=2,dimnames=list(c("e","f"),c())))
merge_list_mx.by.cols(xlistm)
}
}
\seealso{
merge_list_mx.by.rows
}

