% Generated by roxygen2 (4.0.1): do not edit by hand
\name{meta.add}
\alias{meta.add}
\title{Add a meta attribute. Appends to existing meta attribute, if it exists,
overwriting any meta elements with the same name.}
\usage{
meta.add(x, new.meta)
}
\arguments{
\item{x}{an object. May have an existing meta attribute.}

\item{new.meta}{a named vector to add to the meta attribute of x.
 if an element of new.meta already exists in the x's meta,
 then the existing meta element will be replaced.
 if new.meta is NULL, nothing is done.}
}
\value{
x with "meta" attribute
}
\description{
Add a meta attribute. Appends to existing meta attribute, if it exists,
overwriting any meta elements with the same name.
}
\examples{
x <- 1:3
x <- structure(1:3, meta=c(grpby.tag="r1stchildethn", set="females only"))
new.meta <- c(varname="gptotvis")
new.meta <- c(grpby.tag="newtag")
new.meta <- c(grpby.tag=NULL)
meta.add(x, new.meta)
meta.add(x, new.meta)
}

