% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Models.R
\name{model_coefs_names_unique}
\alias{model_coefs_names_unique}
\title{Get the unique set of names of model coefs for the
supplied list of models}
\usage{
model_coefs_names_unique(models_list)
}
\arguments{
\item{models_list}{list of models}
}
\description{
Get the unique set of names of model coefs for the
supplied list of models
}
\examples{
\dontrun{
	model_coefs_names_unique(models)
}
}

