% Generated by roxygen2 (4.0.1): do not edit by hand
\name{namedList}
\alias{namedList}
\title{Create a named list of NULL elements.}
\usage{
namedList(...)
}
\arguments{
\item{...}{names of the elements}
}
\value{
a list of NULL elements
}
\description{
Create a named list of NULL elements.
}
\examples{
namedList("foo","bar")
}

