% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Models.R
\name{predSimNBinomsSelect}
\alias{predSimNBinomsSelect}
\title{Predict and simulate value from n negative binomial models.
No backtransformation included.}
\usage{
predSimNBinomsSelect(x.cat, models, envir = parent.frame())
}
\arguments{
\item{x.cat}{a categorical vector}

\item{models}{a list of models with length equal to the number of categories in x.cat}

\item{envir}{environment in which to evaluate model variables.}
}
\value{
a continuous vector that when binned by cont.bonbreaks will be the same as x.cat
}
\description{
Use when all the catToCont models for a variable are negative binomial.
}
\examples{
\dontrun{}
}

