% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MatrixStats.r
\name{proportions_at_each_run}
\alias{proportions_at_each_run}
\title{Calculates proportions for each iteration for each run.
Called by mean_array_pctile_CIs2().
If xa has a group-by tag then proportions are calculated separately for each group.}
\usage{
proportions_at_each_run(xa, grpby.tag, binbreaks, varname, dict)
}
\arguments{
\item{xa}{an array of run results.  nrow = num iterations,
ncol = num variable groups x num group-by groups, num z dimensions = num runs}

\item{grpby.tag}{NA or the name of the groupby variable}

\item{binbreaks}{Passed from mean_array_pctile_CIS2().  Used to calculate the number of categories for
the outcome variable.}

\item{varname}{a string of length 1 with the name of the outcome variable.  Passed from
mean_array_pctile_CIS2().}

\item{dicr}{the dictionary of the specific MSM project.}
}
\description{
Other way had the means of the numerator collated.  A percentage was calculated by
assuming that the denominator did not change over time.  In this function the group-by
variable can be a time-variant variable and hence the denominator changes over time.
Because of this, the percentages are calculated for each group (if there is grouping),
for each year for each run.  Then later the mean of the percentages is taken rather
than the mean of the frequencies.  For more details see the preamble to
mean_array_pctile_CIs2().
}
\examples{
\dontrun{}
}

