\name{read_csv}
\alias{read_csv}
\title{Reads a CSV file.}
\usage{
  read_csv(filedir, filename, stringsAsFactors = FALSE,
    ...)
}
\arguments{
  \item{filedir}{file directory, with or without trailing
  slash}

  \item{filename}{file name}

  \item{stringsAsFactors}{logical: should character vectors
  be converted to factors?}

  \item{...}{additional parameters to read.csv}
}
\description{
  Reads a CSV file.
}

