\name{readXLSSheet1}
\alias{readXLSSheet1}
\title{Returns first sheet of XLS as dataframe}
\usage{
  readXLSSheet1(filedir, filename)
}
\arguments{
  \item{filedir}{file dir, including trailing slash}

  \item{filename}{file name}
}
\description{
  Returns first sheet of XLS as dataframe
}

