% Generated by roxygen2 (4.0.1): do not edit by hand
\name{remove.cols.named}
\alias{remove.cols.named}
\title{Remove cols by name.}
\usage{
remove.cols.named(x, cnames)
}
\arguments{
\item{x}{matrix or dataframe}

\item{cnames}{vector of colnames}
}
\value{
matrix after removed.
}
\description{
Remove cols by name.
}
\examples{
x <- matrix(1:4, dimnames=list(NULL, c("A","NA (\%)")), nrow = 2)
x <- matrix(1:4, dimnames=list(NULL, c("A","B")), nrow = 2)
cnames <- "NA (\%)"
remove.cols.named(x, cnames)
data(mtcars); x <- cars; cnames <- "dist"
remove.cols.named(x, cnames)
}

