% Generated by roxygen2 (4.0.1): do not edit by hand
\name{remove.NA.cols}
\alias{remove.NA.cols}
\title{Remove columns that contain all NAs.}
\usage{
remove.NA.cols(mx)
}
\arguments{
\item{mx}{matrix}
}
\value{
matrix after removed.
}
\description{
Remove columns that contain all NAs.
}
\examples{
mx <- matrix(c(1:4, NA, NA), nrow=2)
remove.NA.cols(mx)
}

