% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CategoricalAdjustment.r
\name{removeGlobalSubgroupFilterExpression}
\alias{removeGlobalSubgroupFilterExpression}
\title{Clear the subgroup expression for all cat.adjustments.}
\usage{
removeGlobalSubgroupFilterExpression()
}
\description{
Clear the subgroup expression for all cat.adjustments.
}
\examples{
\dontrun{}
}

