% Generated by roxygen2 (4.0.1): do not edit by hand
\name{result.as.means.and.errs}
\alias{result.as.means.and.errs}
\title{Takes a result row and returns the means and error amounts as separate vectors in a matrix or list.}
\usage{
result.as.means.and.errs(result.row, simplify = T)
}
\arguments{
\item{result.row}{a result row, ie: a vector with values named Mean and Lower eg:

>  envs$`Scenario 1`$years1_5$run_results_collated$means$kids["Total",]
    Mean    Lower    Upper
10.99488 10.62256 11.36721

 if there are no values named Mean, then it will be assumed that all values
 are Means and that Lower is 0.}

\item{simplify}{if TRUE (default), returns a matrix instead of a list.}
}
\value{
a matrix/list of means and errs. The first row/means vector is the means from the result row, and the
 second row/errs vector is the difference between each mean and it's lower value.
}
\description{
Takes a result row and returns the means and error amounts as separate vectors in a matrix or list.
}
\examples{
\dontrun{

result.row <- envs$`Scenario 1`$years1_5$run_results_collated$means$kids["Total",]
\dontrun{
> result.row
    Mean    Lower    Upper
10.99488 10.62256 11.36721

> result.as.means.and.errs(result.row)
$means

10.99488

$errs

0.3723213

}
result.row <- c("0\%"=5,"20\%"=5,"40\%"=9,"60\%"=11,"80\%"=15,"100\%"=50)
result.row <- structure(c(5, 5, 5, 5, 5, 5, 9, 9, 9, 11, 11, 11, 15, 15, 15,50.5, 6.02828342338857, 94.9717165766114), .Names = c("0\% Mean","0\% Lower", "0\% Upper", "20\% Mean", "20\% Lower", "20\% Upper","40\% Mean", "40\% Lower", "40\% Upper", "60\% Mean", "60\% Lower","60\% Upper", "80\% Mean", "80\% Lower", "80\% Upper", "100\% Mean","100\% Lower", "100\% Upper"))

result.row <- env.base$modules$years1_5$run_results_collated$quantiles$kids["All Years",]
result.row <- env.scenario$modules$years1_5$run_results_collated$quantiles$kids["All Years",]
result.row <- env.scenario$modules$years1_5$run_results_collated$means$kids["All Years",]
result.row <- na.omit(env.scenario$modules$years6_13$run_results_collated$histogram[["cond"]]["All Years",])

result.as.means.and.errs(result.row.scenario)

result.as.means.and.errs(result.row)
}
}

