\name{save.env}
\alias{save.env}
\title{Convience method for saving environment to
an Rdata file for debugging.}
\usage{
  save.env(env = env.scenario, file = "env.rdata")
}
\arguments{
  \item{env}{environment to save, defaults to env.scenario}

  \item{file}{file name, defaults to "env.rdata"}
}
\description{
  Convience method for saving environment to an Rdata file
  for debugging.
}
\examples{
save.env()
}

