% Generated by roxygen2 (4.0.1): do not edit by hand
\name{setGlobalSubgroupFilterExpression}
\alias{setGlobalSubgroupFilterExpression}
\title{Set the subgroup expression to all cat.adjustments if the subgroup expression exists.
Otherwise, remove the subgroup expression.}
\usage{
setGlobalSubgroupFilterExpression(subgroupExpression)
}
\arguments{
\item{subgroupExpression}{the subgroup expression that is requested by the user.
 It specify the subgroup which is going to be adjusted.}
}
\description{
Set the subgroup expression to all cat.adjustments if the subgroup expression exists.
Otherwise, remove the subgroup expression.
}
\examples{
\dontrun{
subgroupExpression <- "mhrswrk < 20"
setGlobalSubgroupFilterExpression(subgroupExpression)
attr(env.scenario$cat.adjustments[[1]], "logisetexpr")
}
}

