% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/StringUtil.r
\name{strip.alpha.mx}
\alias{strip.alpha.mx}
\title{Remove any alpha from row and col names, leaving only numbers.}
\usage{
strip.alpha.mx(mx)
}
\arguments{
\item{mx}{matrix}
}
\value{
matrix after modify the row and col names
}
\description{
Remove any alpha from row and col names, leaving only numbers.
}
\examples{
mx <- matrix(c(1:4), nrow=2, ncol=2, dimnames=list(c("0 (\%)", "1 (\%)"), c("0 (\%)", "1 (\%)")))
strip.alpha.mx(mx)
}

