% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/StringUtil.r
\name{strip.numeric}
\alias{strip.numeric}
\title{Remove any numbers from character vector, leaving only alpha.}
\usage{
strip.numeric(x)
}
\arguments{
\item{x}{character vector}
}
\value{
character vector without numbers
}
\description{
Remove any numbers from character vector, leaving only alpha.
}
\examples{
x <- c("1 (\%)", "-2 (\%)", "2.102 (\%)", "10 (\%)")
strip.numeric(x)
}

