% Generated by roxygen2 (4.0.1): do not edit by hand
\name{stripEmpty}
\alias{stripEmpty}
\title{Remove empty values from vector}
\usage{
stripEmpty(xvec)
}
\arguments{
\item{xvec}{a vector}
}
\value{
a vector without empty values
}
\description{
Remove empty values from vector
}
\examples{
stripEmpty(c("", NA, "1"))
}

