% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/support.r
\name{stripMeta.list}
\alias{stripMeta.list}
\title{Remove meta attribute from a list}
\usage{
stripMeta.list(xlist)
}
\arguments{
\item{xlist}{list}
}
\value{
xlist with elements without a "meta" attribute
}
\description{
Remove meta attribute from a list
}

