% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/support.r
\name{stripMeta}
\alias{stripMeta}
\title{Remove meta attribute from an object}
\usage{
stripMeta(x)
}
\arguments{
\item{x}{object}
}
\value{
x without "meta" attribute
}
\description{
Remove meta attribute from an object
}

