% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simulate.r
\name{table.catvar.with.CI}
\alias{table.catvar.with.CI}
\title{This is the version for table.catvar() with confidence interval.
Produce a proportioned table for x with confidence interval, using
the specified coding as names and setting the "meta" attribute to "varname"
of coding.}
\usage{
table.catvar.with.CI(x, coding)
}
\arguments{
\item{x}{vector of values}

\item{coding}{a coding variable. names(coding) is the labels
 attr(coding, "varname") is a named element in xlist}
}
\value{
a table (proportions) with names specified by coding
}
\description{
This is the version for table.catvar() with confidence interval.
Produce a proportioned table for x with confidence interval, using
the specified coding as names and setting the "meta" attribute to "varname"
of coding.
}
\examples{
\dontrun{
table.catvar.with.CI(children$SESBTH, codings$SESBTH)
x <- simframe$z1singleLvl1 ; coding <- codings$z1singleLvl1
table.catvar.with.CI(simframe$z1singleLvl1, codings$z1singleLvl1)
x <- binary.levels.combine(simframe.master$SESBTHLvl1, simframe.master$SESBTHLvl2, simframe.master$SESBTHLvl3)
coding <- dict.MELC$codings$SESBTH}
}

