% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simulate.r
\name{table.contvar.with.CI}
\alias{table.contvar.with.CI}
\title{This is the version for table.contvar() with confidence interval.
Display a vector of continuous values in a table with confidence interval
using the breaks supplied.
Attachs a meta attribute with varname}
\usage{
table.contvar.with.CI(x, breaks, varname)
}
\arguments{
\item{x}{vector of continous values}

\item{breaks}{a numeric vector of two or more cut points
NB: note that the cut point value is not included in the bin
(ie: include.lowest = FALSE)
Therefore the very first cut point must be less than min(x)}

\item{varname}{added as a tag on the meta attribute}
}
\value{
a table (proportions) with names specified by breaks
}
\description{
This is the version for table.contvar() with confidence interval.
Display a vector of continuous values in a table with confidence interval
using the breaks supplied.
Attachs a meta attribute with varname
}
\examples{
\dontrun{
x <- env.scenario$simframe$bwkg
breaks <- binbreaks$bwkg

table.contvar.with.CI(env.scenario$simframe$bwkg, binbreaks$bwkg, "bwkg")
}
}

