% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MatrixStats.r
\name{table.grpby_BCASO1}
\alias{table.grpby_BCASO1}
\title{Frequency table, with option to group results.
Extension of table.grpby() that can handle grpby as a matrix as well as a vector.
The group-by variable may be time-invariant or time-variant.}
\usage{
table.grpby_BCASO1(x, grpby = NULL, wgts = NULL, binbreak = NULL)
}
\arguments{
\item{x}{vector of values which can be interpreted as factors}

\item{grpby}{A vector or matrix of elements to group by, or NULL to do no grouping.
 vector must be same length as the columns of x or matrix must have the same number
 of rows as z.
If the group-by variable is not time-invariant grpby can be provided as a vector or as a
matrix with every column the same.  If the group-by variable is time-variable then
grpby should be a matrix with number of columns equal to number of years.}

\item{wgts}{vector of weights, or NULL to do no weighting
 Same length as the columns of x.}

\item{binbreak}{The binbreaks for the specific outcome variable.}
}
\value{
a table. If grpby is specified this will be a table
 with columns that are the group by and rows the categories.
 If grpby = NULL then a table with 1 column and rows as categories is returned.
}
\description{
Frequency table, with option to group results.
Extension of table.grpby() that can handle grpby as a matrix as well as a vector.
The group-by variable may be time-invariant or time-variant.
}
\examples{
\dontrun{}
}

