\name{unique.names.model.coefs}
\alias{unique.names.model.coefs}
\title{Get the unique set of names of model coefs for the
supplied list of models}
\usage{
  unique.names.model.coefs(models_list)
}
\arguments{
  \item{models_list}{list of models}
}
\description{
  Get the unique set of names of model coefs for the
  supplied list of models
}
\examples{
\dontrun{
	unique.names.model.coefs(models)
}
}

