\name{updatelist}
\alias{updatelist}
\title{Update dest with the values in src, removing any values from dest that don't exist in src.}
\usage{
  updatelist(dest, src)
}
\arguments{
  \item{dest}{dest list}

  \item{src}{src list}
}
\description{
  Use this to update a dest list when the ordering of src
  is different from dest and you wwsh to preserve the dest
  list ordering.
}

