% Generated by roxygen2 (4.0.1): do not edit by hand
\name{as.arrayFromList}
\alias{as.arrayFromList}
\title{convert list of vectors to an array.
Recycling rule will be applied.}
\usage{
as.arrayFromList(mylist)
}
\arguments{
\item{mylist}{list of vectors}
}
\value{
an array
}
\description{
convert list of vectors to an array.
Recycling rule will be applied.
}
\examples{
mylist <- list(1:5,2:6,7)
result <- as.arrayFromList(mylist)
}

