\name{as.csv.string.list}
\alias{as.csv.string.list}
\title{Apply to a list using the names from the list}
\usage{
  as.csv.string.list(xlist, row.names = T)
}
\arguments{
  \item{xlist}{list}

  \item{row.names}{either a logical value indicating
  whether the row names of x are to be written along with
  x, or a character vector of row names to be written}
}
\description{
  Apply to a list using the names from the list
}
\examples{
#cat(as.csv.string.list(t.runstats.mean.cfreq.base))
}

