% Generated by roxygen2 (4.0.1): do not edit by hand
\name{as_data_frame_list_as_is}
\alias{as_data_frame_list_as_is}
\title{Convert list to dataframe simply by setting class attr
This preserves any matrices in the list.}
\usage{
as_data_frame_list_as_is(xlist, row.names = NULL)
}
\arguments{
\item{xlist}{list}

\item{row.names}{row names for the data.frame}
}
\value{
a data frame
}
\description{
Convert list to dataframe simply by setting class attr
This preserves any matrices in the list.
}
\examples{
x <- list(matrix(1:4, nrow=2),c(1:4))
x <- as_data_frame_list_as_is(x)
}

