% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/support.r
\name{assert}
\alias{assert}
\title{Stops and displays indices of any element that is FALSE.}
\usage{
assert(xvec)
}
\arguments{
\item{xvec}{logical vector}
}
\value{
If there is FALSE, return error message.
}
\description{
Stops and displays indices of any element that is FALSE.
}
\examples{
assert(c(TRUE,TRUE))
#assert(c(TRUE,FALSE))
}

