\name{checkModelVars}
\alias{checkModelVars}
\title{Checks that the variables used in all of the models
exist in simvalues, if not generate stop}
\usage{
  checkModelVars(models, simframe)
}
\arguments{
  \item{models}{a list of glm objects}

  \item{simframe}{the simframe}
}
\description{
  Checks that the variables used in all of the models exist
  in simvalues, if not generate stop
}
\examples{
#simframe <- simframe.start
#checkModelVars(models, simframe)
}

