% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ListUtil.r
\name{checkNAs}
\alias{checkNAs}
\title{Check a list for any NAs, producing error if they exist, otherwise
silently exit.}
\usage{
checkNAs(xlist)
}
\arguments{
\item{xlist}{list to check}
}
\value{
nothing if no NAs, otherwise error message
}
\description{
Check a list for any NAs, producing error if they exist, otherwise
silently exit.
}
\examples{
\dontrun{
	foo <- c(1,NA,3)
 names(foo) <- c("a","b","c")
 xlist <- foo

	checkNAs(foo)
	checkNAs(mpropens)
}
}

