% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Collate.r
\name{collator_histogram}
\alias{collator_histogram}
\title{Collates frequencies for use in histogram output with confidence intervals.
Performs the following:}
\usage{
collator_histogram(runs, dict, row.dim.label = "Year", col.dim.label = "")
}
\arguments{
\item{runs}{a list of lists of matrices, one inner list per run.
 Each inner list may have any number of matrices,
 and each matrix may have a different sets of rows or columns.
 The matrices will be flattened into rows.}

\item{dict}{Dictionary object. Used to label columns.}

\item{row.dim.label}{name of the entire row dimension}

\item{col.dim.label}{name of the entire col dimension}
}
\description{
\itemize{
  \item Takes mean with confidence intervals using \code{\link{collator_mutiple_lists_mx}}
  \item Labels the result using the dictionary
}
}
\seealso{
\code{\link{collator_mutiple_lists_mx}}
}

