% Generated by roxygen2 (4.0.1): do not edit by hand
\name{collator_mutiple_lists_mx2}
\alias{collator_mutiple_lists_mx2}
\title{A new version of collator_mutiple_lists_mx() that calls
mean_array_pctile_CIs2()) instead of then 'non-2' version.  This means that
percenatages will be averaged over runs instead of frequencies and the confidence
intervals will be calculated on percentages rather than frequencies.  See documentation
on mean_array_pctile_CIS2() more more details.
Another difference between this function and the original collator_mutiple_lists_mx()
is that this function takes cat.adjustments, dict, and binbreaks arguments.}
\usage{
collator_mutiple_lists_mx2(runs, CI = TRUE, cat.adjustments = NULL, dict,
  binbreaks = NULL)
}
\arguments{
\item{runs}{a list of lists of matrices, one inner list per run.
 Each inner list may have any number of matrices,
 and each matrix may have a different sets of rows or columns.
 The matrices will be flattened into rows.}

\item{CI}{if TRUE and length(runs) > 1, lower and upper confidence intervals
 are returned in additional columns}

\item{cat.adjustments}{The cat.adjustments list containing the cat.adjustments for multiple variables.
 Within the function the specific cat.adjustments for the variable of interest are
 extracted from the list in further embedded functions.}

\item{dict}{Dictionary object.}

\item{binbreaks}{The binbreaks for the specific outcome variable.}
}
\value{
a matrix with the averaged values of runs.
}
\description{
A new version of collator_mutiple_lists_mx() that calls
mean_array_pctile_CIs2()) instead of then 'non-2' version.  This means that
percenatages will be averaged over runs instead of frequencies and the confidence
intervals will be calculated on percentages rather than frequencies.  See documentation
on mean_array_pctile_CIS2() more more details.
Another difference between this function and the original collator_mutiple_lists_mx()
is that this function takes cat.adjustments, dict, and binbreaks arguments.
}
\examples{
\dontrun{}
}
\keyword{internal}

