\name{detachReturn}
\alias{detachReturn}
\title{Detach returning the modified attached environment.
NB: the returned environment contains the modified
    contents of the attached environment but is
    actually a newly created different environment
    object from the original.}
\usage{
  detachReturn(envname)
}
\description{
  Detach returning the modified attached environment. NB:
  the returned environment contains the modified contents
  of the attached environment but is actually a newly
  created different environment object from the original.
}

