% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{dict_example}
\alias{dict_example}
\title{An example Dictionary proto object.}
\format{A Dictionary proto object}
\usage{
dict_example
}
\description{
Provides descriptions for
\itemize{
  \item age
  \item age_grp
  \item alive
  \item disability_state
  \item earnings
  \item sex
}
}
\details{
Provides codings for
\itemize{
  \item age_grp
  \item alive
  \item disability_state
  \item sex
}
}
\keyword{datasets}

