% Generated by roxygen2 (4.0.1): do not edit by hand
\name{evaluateLogisetExprAttribute}
\alias{evaluateLogisetExprAttribute}
\title{Evaluate logiset expression attribute for the variable.}
\usage{
evaluateLogisetExprAttribute(desired_props, simframe, varname = "")
}
\arguments{
\item{desired_props}{a vector that is the proportions requested by the user.
The vector is the length of the number of distinct values of the variable
being modified.}

\item{simframe}{the simframe to evaluate}

\item{varname}{variable name.}
}
\value{
a vector of logical value
}
\description{
Evaluate logiset expression attribute for the variable.
}
\examples{
\dontrun{
desired_props <- structure(1, logiset="alive & residential")
simframe <- env.base$simframe
evaluateLogisetExprAttribute(desired_props, simframe)
}
}

