\name{global}
\alias{global}
\title{Save x into global variable as it's own name.}
\usage{
  global(x, pos = 1)
}
\arguments{
  \item{x}{value}

  \item{pos}{environment to save into. Defaults to global
  environment.}
}
\description{
  Save x into global variable as it's own name.
}

