% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/support.r
\name{globalNamed}
\alias{globalNamed}
\title{Save x into global variable, ie: top frame, not just this function
using the supplied varname}
\usage{
globalNamed(varname, x, pos = 1)
}
\arguments{
\item{varname}{variable name}

\item{x}{value}

\item{pos}{environment to save into. Defaults to global environment.}
}
\description{
Save x into global variable, ie: top frame, not just this function
using the supplied varname
}

