% Generated by roxygen2 (4.0.1): do not edit by hand
\name{isListOfNulls}
\alias{isListOfNulls}
\title{True if all list members are NULL}
\usage{
isListOfNulls(xlist)
}
\arguments{
\item{xlist}{list}
}
\value{
a logical value
}
\description{
True if all list members are NULL
}
\examples{
x <- list(NULL)
isListOfNulls(x)
}

