% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Label.r
\name{labelColTitleFromList}
\alias{labelColTitleFromList}
\title{names each object's column title with the
object's name in the named list}
\usage{
labelColTitleFromList(xnamedlist)
}
\arguments{
\item{xnamedlist}{list}
}
\value{
a named list with named column
}
\description{
names each object's column title with the
object's name in the named list
}

