% Generated by roxygen2 (4.0.1): do not edit by hand
\name{loadGLMXLS}
\alias{loadGLMXLS}
\title{Load and create a GLM from the first sheet in an xlsx file.
See \code{\link{createGLM}} for file format.}
\usage{
loadGLMXLS(filedir, filename)
}
\arguments{
\item{filedir}{file directory, ending with "/", eg: "d:/workspace/"}

\item{filename}{file name, eg: "myfile.xls". See \code{\link{createGLM}} for file format.}
}
\value{
a glm model object
}
\description{
Load and create a GLM from the first sheet in an xlsx file.
See \code{\link{createGLM}} for file format.
}
\examples{
\dontrun{
filedir <- "D:/workspace.sim/MELC/CHDS/models/"
filename <- "gptotvis.xlsx" ; filename <- "paddhsbetas.xlsx"

loadGLMXLS(filedir, filename)
}
}
\seealso{
See \code{\link{createGLM}} for file format.
}

