\name{merge.list.mx.by.cols}
\alias{merge.list.mx.by.cols}
\title{Takes any number of 2D matrices, each with the same number of cols but with
any set of named rows. Every col 1 of the matrices are combined into a new
matrix, as are every col 2,3 etc.}
\usage{
  merge.list.mx.by.cols(...)
}
\arguments{
  \item{...}{matrices, or a list of matrices}
}
\description{
  Takes any number of 2D matrices, each with the same
  number of cols but with any set of named rows. Every col
  1 of the matrices are combined into a new matrix, as are
  every col 2,3 etc.
}
\seealso{
  merge.list.mx.by.rows
}

