% Generated by roxygen2 (4.0.1): do not edit by hand
\name{nsort}
\alias{nsort}
\title{Sort numerically.
If x can be represented numerically, then
sort numerically otherwise sort normally.}
\usage{
nsort(x, stripAlpha = TRUE, sortAlphaOnlySeparately = TRUE, ...)
}
\arguments{
\item{x}{object to sort}

\item{stripAlpha}{remove alpha characters before attempting sort, but retain them in the output}

\item{sortAlphaOnlySeparately}{if a value is purely alpha (ie: contains no numeric component) sort is separately
 from the alpha-numeric values, and return it after any numerically sorted values}

\item{...}{additional arguments passed to sort}
}
\value{
x after sorted.
}
\description{
Sort numerically.
If x can be represented numerically, then
sort numerically otherwise sort normally.
}
\examples{
x <- c("1 (\%)", "-2 (\%)", "2 (\%)", "10 (\%)")
x <- c("1 (\%)", "0 (\%)", "2 (\%)", "10 (\%)")
x <- c("1","0","2","10")
x <- c("b","d","a","c")
x <- c(1,4,3,2)
x <- c("1", "2", "3", "4", "5", "6", "7", "8", "10", NA, "9", NA)
x <- c("10", "11", "12", "13", "6", "7", "8", "9", "All Years")
x <- c("10 (\%)", "11 (\%)", "12", "13", "6", "7", "8", "9", "All Years")
nsort(x)
}
\seealso{
sort
}

