% Generated by roxygen2 (4.0.1): do not edit by hand
\name{predSimBinom}
\alias{predSimBinom}
\title{Predict and simulate binary value from binomial
distribution with probability}
\usage{
predSimBinom(model.glm, envir = parent.frame(), set = NULL)
}
\arguments{
\item{model.glm}{model specifying variables to evaluate and coefficients
to multiple by.}

\item{envir}{environment in which to evaluate model variables.
if unspecified, uses caller's environment}

\item{set}{logical vector indicating elements or rows to simulate, or NULL to
 simulate using all values in envir}
}
\value{
a vector of binary values drawn from a Binomial distribution
}
\description{
Predict and simulate binary value from binomial
distribution with probability
}
\examples{
\dontrun{
model.glm <- models$houtptot
newdata <- simvalues
}
}

