% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Models.R
\name{predSimNBinom}
\alias{predSimNBinom}
\title{Predict and simulate continuous value from negative binomial
distribution with mean = exp(predicted) and size=1/alpha
where alpha is specified in the model.}
\usage{
predSimNBinom(model.glm, envir = parent.frame(), set = NULL, alpha = NULL)
}
\arguments{
\item{model.glm}{model specifying variables to evaluate and coefficients
to multiple by.}

\item{envir}{environment in which to evaluate model variables.
if unspecified, uses caller's environment}

\item{set}{logical vector indicating elements or rows to simulate, or NULL to
simulate using all values in envir}

\item{alpha}{if supplied, use this value for alpha rather than the value in the model}
}
\value{
a vector of predicted value
}
\description{
Predict and simulate continuous value from negative binomial
distribution with mean = exp(predicted) and size=1/alpha
where alpha is specified in the model.
}
\examples{
\dontrun{
model.glm <- models$houtptot
newdata <- simvalues
predSimNBinom (model.glm, newdata)
}
}

