% Generated by roxygen2 (4.0.1): do not edit by hand
\name{predSimPois}
\alias{predSimPois}
\title{Predict and simulate continuous value from poisson model}
\usage{
predSimPois(model.glm, envir = parent.frame(), set = NULL)
}
\arguments{
\item{model.glm}{model specifying variables to evaluate and coefficients
to multiple by.}

\item{envir}{environment in which to evaluate model variables.
if unspecified, uses caller's environment}

\item{set}{logical vector indicating elements or rows to simulate, or NULL to
 simulate using all values in envir}
}
\value{
a vector of predicted value
}
\description{
Predict and simulate continuous value from poisson model
}
\examples{
\dontrun{
 model.glm <- models$hadmtot
	newdata <- simvalues
}
}

