% Generated by roxygen2 (4.0.1): do not edit by hand
\name{read_csv}
\alias{read_csv}
\title{Reads a CSV file.}
\usage{
read_csv(filedir, filename, stringsAsFactors = FALSE, ...)
}
\arguments{
\item{filedir}{file directory, with or without trailing slash}

\item{filename}{file name}

\item{stringsAsFactors}{logical: should character vectors be converted to factors?}

\item{...}{additional parameters to read.csv}
}
\value{
a data frame
}
\description{
Reads a CSV file.
}
\examples{
\dontrun{
filedir <- "D:/workspace.sim/simario/demo/data/base"
people <<- read_csv(base_dir, "Base_file_(people).csv")
}
}

