% Generated by roxygen2 (4.0.1): do not edit by hand
\name{readXLSSheet1}
\alias{readXLSSheet1}
\title{Returns first sheet of XLS as dataframe}
\usage{
readXLSSheet1(filedir, filename, stringsAsFactors = FALSE, ...)
}
\arguments{
\item{filedir}{file directory, with or without trailing slash}

\item{filename}{file name}

\item{stringsAsFactors}{logical: should character vectors be converted to factors?}

\item{...}{additional parameters to read.xlsx2}
}
\value{
a data frame
}
\description{
Returns first sheet of XLS as dataframe
}
\examples{
\dontrun{}
}

