% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MatrixUtil.r
\name{remove.cols}
\alias{remove.cols}
\title{Remove cols specified by indices.}
\usage{
remove.cols(mx, indices)
}
\arguments{
\item{mx}{matrix}

\item{indices}{indices to remove. If indices is length zero, then mx is return as is.}
}
\value{
the matrix mx with columns removed. Names and any "meta" attribute preserved.
 Always returns a matrix.
}
\description{
Remove cols specified by indices.
}
\examples{
mx <- matrix(1:3, nrow=1, dimnames=list(NULL, c("A","B","C")))
 indices <- c(1,3)
 indices <- vector(mode="integer")
 remove.cols(mx, indices)
}

