% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MatrixUtil.r
\name{remove.zero.cols}
\alias{remove.zero.cols}
\title{Remove cols that contain all zeros.}
\usage{
remove.zero.cols(mx)
}
\arguments{
\item{mx}{matrix}
}
\value{
matrix after removed.
}
\description{
Remove cols that contain all zeros.
}
\examples{
mx <- matrix(c(1:4, 0, 0), nrow=2)
remove.zero.cols(mx)
}

