% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{Simenv}
\alias{Simenv}
\title{Simenv object - a simulation environment.}
\format{\preformatted{proto object 
 $ class                                        :function (.)   
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 95 11 95 30 18 37 95 95 
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x000000001ce14570>  
 $ applyCatAdjustmentToSimframeVarMultipleBinary:function (., binLevelVarnames, desiredProps, propens, printAdj = TRUE,  
    logiset = NULL)   
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 409 51 485 2 58 9 409 485 
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x000000001ce14570>  
 $ applyContAdjustmentToSimframe                :function (., varname, iteration, desiredProps, catToContModels, cont.binbreaks,  
    propensities)   
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 522 35 528 2 42 9 522 528 
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x000000001ce14570>  
 $ numberOfUnits                                :function (.)   
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 627 19 629 2 26 9 627 629 
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x000000001ce14570>  
 $ clone                                        :function (.)   
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 90 11 90 57 18 64 90 90 
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x000000001ce14570>  
 $ applyAllFixedOutcomesIfSetToSimframe         :function (.)   
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 210 42 218 2 49 9 210 218 
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x000000001ce14570>  
 $ simulate                                     :function (., total_runs = 1)   
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 562 14 624 2 21 9 562 624 
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x000000001ce14570>  
 $ generatePreSimulationStats                   :function (., simframe)   
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 537 32 539 2 39 9 537 539 
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x000000001ce14570>  
 $ applyAllCatAdjustmentsToSimframe             :function (., iteration, propensities = NULL, print_adj = TRUE, cat.adjustments = .$cat.adjustments)   
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 164 38 206 2 45 9 164 206 
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x000000001ce14570>  
 $ applyCatAdjustmentToSimframeVarSingle        :function (., varname, desired_props, propens, print_adj = T, logiset = NULL)   
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 331 43 353 2 50 9 331 353 
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x000000001ce14570>  
 $ new                                          :function (., name, simframe, dict, cat.adjustments = list(), modules = list())   
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 74 9 85 2 16 9 74 85 
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x000000001ce14570>  
 $ applyCatAdjustmentToSimframe                 :function (., varnames, desired_props, iteration, propensities, print_adj = TRUE)   
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 274 34 287 2 41 9 274 287 
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x000000001ce14570>  
}}
\usage{
Simenv
}
\description{
A simulation environment contains everything required to perform a simulation. Typically 1 Simenv will be created
and used to run a base simulation, and additional Simenvs will be created to test different scenarios.
}
\details{
A Simenv consists of a:
- a simframe (possibly with adjustments to test a scenario)
- one or more simulation modules (Simmodule). A Simmodule contains outcomes, run stats, and runs.averaged for a simulation
  as well as the code to generate them.

Uses the global environment list variable "propensities" when performing categorical adjustment

This class will be subclassed by specific simulation problems which will provide their own simframe,
Simmodules and adjustments.
}
\keyword{datasets}

