% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Simenv.r
\docType{data}
\name{Simenv}
\alias{Simenv}
\title{Simenv object - a simulation environment.}
\format{\preformatted{proto object 
 $ class                                        :function (.)   
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 95 11 95 30 18 37 95 95 
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x0000000009d81f70>  
 $ simulateP                                    :function (., total_runs = 1)   
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 681 15 746 2 22 9 681 746 
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x0000000009d81f70>  
 $ check.subgroup.expr                          :function (., cat.adjustments = .$cat.adjustments, simframe = .$simframe)   
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 550 25 565 2 32 9 550 565 
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x0000000009d81f70>  
 $ applyCatAdjustmentToSimframeVarMultipleBinary:function (., binLevelVarnames, desiredProps, propens, printAdj = TRUE, logiset = NULL)   
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 418 51 494 2 58 9 418 494 
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x0000000009d81f70>  
 $ applyContAdjustmentToSimframe                :function (., varname, iteration, desiredProps, catToContModels, cont.binbreaks, propensities)   
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 531 35 537 2 42 9 531 537 
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x0000000009d81f70>  
 $ numberOfUnits                                :function (.)   
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 750 19 752 2 26 9 750 752 
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x0000000009d81f70>  
 $ clone                                        :function (.)   
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 90 11 90 57 18 64 90 90 
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x0000000009d81f70>  
 $ applyAllFixedOutcomesIfSetToSimframe         :function (.)   
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 215 42 223 2 49 9 215 223 
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x0000000009d81f70>  
 $ simulate                                     :function (., total_runs = 1)   
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 588 14 658 2 21 9 588 658 
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x0000000009d81f70>  
 $ generatePreSimulationStats                   :function (., simframe)   
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 546 32 548 2 39 9 546 548 
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x0000000009d81f70>  
 $ applyAllCatAdjustmentsToSimframe             :function (., iteration, propensities = NULL, print_adj = TRUE, cat.adjustments = .$cat.adjustments)   
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 164 38 209 2 45 9 164 209 
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x0000000009d81f70>  
 $ applyCatAdjustmentToSimframeVarSingle        :function (., varname, desired_props, propens, print_adj = T, logiset = NULL)   
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 340 43 362 2 50 9 340 362 
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x0000000009d81f70>  
 $ new                                          :function (., name, simframe, dict, cat.adjustments = list(), modules = list())   
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 74 9 85 2 16 9 74 85 
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x0000000009d81f70>  
 $ applyCatAdjustmentToSimframe                 :function (., varnames, desired_props, iteration, propensities, print_adj = TRUE)   
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 283 34 296 2 41 9 283 296 
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x0000000009d81f70>  
}}
\usage{
Simenv
}
\description{
A simulation environment contains everything required to perform a simulation. Typically 1 Simenv will be created
and used to run a base simulation, and additional Simenvs will be created to test different scenarios.
}
\details{
A Simenv consists of a:
- a simframe (possibly with adjustments to test a scenario)
- one or more simulation modules (Simmodule). A Simmodule contains outcomes, run stats, and runs.averaged for a simulation
  as well as the code to generate them.

Uses the global environment list variable "propensities" when performing categorical adjustment

This class will be subclassed by specific simulation problems which will provide their own simframe,
Simmodules and adjustments.
}
\keyword{datasets}

