% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{Simenv}
\alias{Simenv}
\title{Simenv object - a simulation environment.}
\format{\preformatted{proto object 
 $ class                                        :function (.)   
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 95 11 95 30 18 37 95 95 
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x0000000013a6a2b8>  
 $ applyCatAdjustmentToSimframeVarMultipleBinary:function (., binLevelVarnames, desiredProps, propens, printAdj = TRUE,  
    logiset = NULL)   
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 414 51 490 2 58 9 414 490 
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x0000000013a6a2b8>  
 $ applyContAdjustmentToSimframe                :function (., varname, iteration, desiredProps, catToContModels, cont.binbreaks,  
    propensities)   
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 527 35 533 2 42 9 527 533 
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x0000000013a6a2b8>  
 $ numberOfUnits                                :function (.)   
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 632 19 634 2 26 9 632 634 
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x0000000013a6a2b8>  
 $ clone                                        :function (.)   
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 90 11 90 57 18 64 90 90 
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x0000000013a6a2b8>  
 $ applyAllFixedOutcomesIfSetToSimframe         :function (.)   
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 215 42 223 2 49 9 215 223 
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x0000000013a6a2b8>  
 $ simulate                                     :function (., total_runs = 1)   
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 567 14 629 2 21 9 567 629 
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x0000000013a6a2b8>  
 $ generatePreSimulationStats                   :function (., simframe)   
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 542 32 544 2 39 9 542 544 
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x0000000013a6a2b8>  
 $ applyAllCatAdjustmentsToSimframe             :function (., iteration, propensities = NULL, print_adj = TRUE, cat.adjustments = .$cat.adjustments)   
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 164 38 211 2 45 9 164 211 
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x0000000013a6a2b8>  
 $ applyCatAdjustmentToSimframeVarSingle        :function (., varname, desired_props, propens, print_adj = T, logiset = NULL)   
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 336 43 358 2 50 9 336 358 
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x0000000013a6a2b8>  
 $ new                                          :function (., name, simframe, dict, cat.adjustments = list(), modules = list())   
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 74 9 85 2 16 9 74 85 
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x0000000013a6a2b8>  
 $ applyCatAdjustmentToSimframe                 :function (., varnames, desired_props, iteration, propensities, print_adj = TRUE)   
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 279 34 292 2 41 9 279 292 
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x0000000013a6a2b8>  
}}
\usage{
Simenv
}
\description{
A simulation environment contains everything required to perform a simulation. Typically 1 Simenv will be created
and used to run a base simulation, and additional Simenvs will be created to test different scenarios.
}
\details{
A Simenv consists of a:
- a simframe (possibly with adjustments to test a scenario)
- one or more simulation modules (Simmodule). A Simmodule contains outcomes, run stats, and runs.averaged for a simulation
  as well as the code to generate them.

Uses the global environment list variable "propensities" when performing categorical adjustment

This class will be subclassed by specific simulation problems which will provide their own simframe,
Simmodules and adjustments.
}
\keyword{datasets}

