% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Simmodule.r
\docType{data}
\name{Simmodule}
\alias{Simmodule}
\title{Simulation module object.}
\format{\preformatted{proto object 
 $ class                      :function (.)   
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 123 11 123 33 18 40 123 123 
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x0000000009cff930>  
 $ clone                      :function (.)   
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 119 11 119 57 18 64 119 119 
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x0000000009cff930>  
 $ collate_all_run_results    :function (., all_run_results)   
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 112 29 114 2 36 9 112 114 
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x0000000009cff930>  
 $ simulateRun                :function (., simenv)   
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 69 17 71 2 24 9 69 71 
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x0000000009cff930>  
 $ new                        :function (., name)   
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 42 9 50 2 16 9 42 50 
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x0000000009cff930>  
 $ map_outcomes_to_run_results:function (., simframe, outcomes)   
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 93 33 95 2 40 9 93 95 
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x0000000009cff930>  
}}
\usage{
Simmodule
}
\description{
A simulation module is really the core of a simulation. It contains the code and output for a distinct set
of results generated, eg: health outcomes for years 1 - 10.
}
\details{
It contains the following key elements:

outcomes - a list of all outcome matrices for the Simmodule.

each Simmodule has a \code{simulateRun} method which transforms the simframe. Typically, transformations will
move variables for micro-units in the simframe through multiple iterations (or time steps).
At the end of each iteration, for outcome variables (as defined in the simframe), the current values
for all micro-units are stored in an outcome matrix.
An outcome matrix contains the set of values for each micro-unit during each iteration.

At the end of each run a set of run stats is calculated for outcomes. A run stat is essentially a function that takes
an outcome matrix and produces an aggregate value for each iteration.
This aggregate value may be a single value (eg: mean), a vector (eg: frequencies, quantiles, summary),
or a matrix (eg: 2 way table).

Run stats are averaged across multiple runs by collateRunStats to get a final simulation result.
}
\keyword{datasets}

