% Generated by roxygen2 (4.0.1): do not edit by hand
\name{strip.alpha}
\alias{strip.alpha}
\title{Remove any alpha from character vector, leaving only numbers.}
\usage{
strip.alpha(x)
}
\arguments{
\item{x}{character vector}
}
\value{
character vector without alpha
}
\description{
Remove any alpha from character vector, leaving only numbers.
}
\examples{
x <- c("1 (\%)", "-2 (\%)", "2.102 (\%)", "10 (\%)")
strip.alpha(x)
}

