% Generated by roxygen2 (4.0.1): do not edit by hand
\name{strip_lvl_suffix}
\alias{strip_lvl_suffix}
\title{Remove trailing "LvlX" (if any) where X is any character}
\usage{
strip_lvl_suffix(varname)
}
\arguments{
\item{varname}{character vector to strip}
}
\value{
a vector of characters without "LvlX"
}
\description{
Remove trailing "LvlX" (if any) where X is any character
}
\examples{
\dontrun{
varname <- c("fooLvl1", "bar")
strip_lvl_suffix(varname)
}
}

