% Generated by roxygen2 (4.0.1): do not edit by hand
\name{stripClass}
\alias{stripClass}
\title{Remove class attribute from an object}
\usage{
stripClass(x)
}
\arguments{
\item{x}{object}
}
\value{
x without "class" attribute
}
\description{
Remove class attribute from an object
}
\examples{
x <- 1:10
x <- structure(x, class=class(x))
stripClass(x)
}

