% Generated by roxygen2 (4.0.1): do not edit by hand
\name{stripMeta}
\alias{stripMeta}
\title{Remove meta attribute from an object}
\usage{
stripMeta(x)
}
\arguments{
\item{x}{object}
}
\value{
x without "meta" attribute
}
\description{
Remove meta attribute from an object
}
\examples{
\dontrun{}
}

