% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MatrixStats.r
\name{table.grpby}
\alias{table.grpby}
\title{Frequency table, with option to group results.
The group-by variable must be time-invariant.}
\usage{
table.grpby(x, grpby = NULL, useNA = "ifany")
}
\arguments{
\item{x}{vector of values which can be interpreted as factors}

\item{grpby}{vector of elements to group by, or NULL to do no grouping
Same length as the columns of x.}

\item{useNA}{whether to include extra NA levels in the table.
 one of c("no", "ifany", "always"). Defaults to "ifany".}
}
\value{
a table. If grpby is specified this will be a table
 with columns that are the group by and rows the categories.
 If grpby = NULL then a table with 1 column and rows as categories is returned.
}
\description{
Frequency table, with option to group results.
The group-by variable must be time-invariant.
}
\examples{
\dontrun{
x <- rep(0,1075)
x <- c(8,8,2,1,1,8)
grpby <- c('M','M','F','F','F','F')

table.grpby(x)
table.grpby(x, grpby)}
}

