% Generated by roxygen2 (4.0.1): do not edit by hand
\name{updatelist}
\alias{updatelist}
\title{Update dest with the values in src, removing any values from dest that don't exist in src.}
\usage{
updatelist(dest, src)
}
\arguments{
\item{dest}{dest list}

\item{src}{src list}
}
\value{
the list "dest" after updated.
}
\description{
Use this to update a dest list when the ordering of src is different from dest and you
wish to preserve the dest list ordering.
}
\examples{
dest <- list(a=1,b=2,c=3,d=4)
src <- list(c=3,a=1,e=5)
updatelist(dest, src)
}

