\name{updatelist}
\alias{updatelist}
\title{Update dest with the values in src, removing any values from dest that don't exist in src.}
\usage{
  updatelist(dest, src)
}
\description{
  Use this to update a dest list when the ordering of src
  is different from dest and you wish to preserve the dest
  list ordering.
}

